tg-2817.[P1] Error 500 al suspender o finalizar medicamento habitual

+++ back-end/hospital-api/src/main/java/net/pladema/clinichistory/requests/medicationrequests/controller/dto/ChangeStateMedicationRequestDto.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class ChangeStateMedicationRequestDto implements Serializable {
 
     @Nullable
-    private Short dayQuantity;
+    private Double dayQuantity;
 
     @Nullable
     private String observations;
