+++ back-end/hospital-api/src/test/java/net/pladema/clinichistory/requests/medicationrequests/service/impl/ChangeStateMedicationServiceImplTest.java
@@ -64,14 +64,14 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
     @Test
     public void test_execute_withInvalidPatientInfo(){
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(null, new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, (short)2, null, Collections.emptyList()))
+                changeStateMedicationService.execute(null, new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, 2d, null, Collections.emptyList()))
         );
         String expectedMessage = "La información del paciente es obligatoria";
         String actualMessage = exception.getMessage();
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(null, (short)1, (short)4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, (short)2, null, Collections.emptyList()))
+                changeStateMedicationService.execute(new PatientInfoBo(null, (short)1, (short)4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, 2d, null, Collections.emptyList()))
         );
         expectedMessage = "El código identificador del paciente es obligatorio";
         actualMessage = exception.getMessage();
@@ -81,14 +81,14 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
     @Test
     public void test_execute_withInvalidNewStatus() {
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(1, (short)1, (short)4), new ChangeStateMedicationRequestBo(null, (short)2, null, Collections.emptyList()))
+                changeStateMedicationService.execute(new PatientInfoBo(1, (short)1, (short)4), new ChangeStateMedicationRequestBo(null, 2d, null, Collections.emptyList()))
         );
         String expectedMessage = "El estado es obligatorio";
         String actualMessage = exception.getMessage();
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(1, (short)1, (short)4),  new ChangeStateMedicationRequestBo("ESTADO", (short)2, null, Collections.emptyList()))
+                changeStateMedicationService.execute(new PatientInfoBo(1, (short)1, (short)4),  new ChangeStateMedicationRequestBo("ESTADO", 2d, null, Collections.emptyList()))
         );
         expectedMessage = "El estado de la medicación es invalido";
         actualMessage = exception.getMessage();
@@ -108,7 +108,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medicationId = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.STOPPED, null, 9, dosageId)).getId();
 
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4),  new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, (short)2, null, List.of(medicationId)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4),  new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, 2d, null, List.of(medicationId)))
         );
         String expectedMessage = "La medicación con id "+ medicationId + " no se puede suspender porque ya esta finalizada";
         String actualMessage = exception.getMessage();
@@ -118,7 +118,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medication2Id = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.STOPPED, null, 9, null)).getId();
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4),  new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, (short)2, null, List.of(medication2Id)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4),  new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, 2d, null, List.of(medication2Id)))
         );
         expectedMessage = "La medicación con id "+ medication2Id + " no se puede suspender porque ya esta finalizada";
         actualMessage = exception.getMessage();
@@ -127,7 +127,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medication3Id = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.SUSPENDED, null, 9, null)).getId();
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, (short)2, null, List.of(medication3Id)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, 2d, null, List.of(medication3Id)))
         );
         expectedMessage = "La medicación con id "+ medication3Id + " no se puede suspender porque ya esta suspendida";
         actualMessage = exception.getMessage();
@@ -140,7 +140,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medication4Id = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.SUSPENDED, null, 9, dosageId)).getId();
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, (short)2, null, List.of(medication4Id)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short)1, (short)4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.SUSPENDED, 2d, null, List.of(medication4Id)))
         );
         expectedMessage = "La medicación con id "+ medication4Id + " no se puede suspender porque ya esta suspendida";
         actualMessage = exception.getMessage();
@@ -166,7 +166,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medicationId = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.STOPPED, null, 9, dosageId)).getId();
 
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, (short)2, null, List.of(medicationId)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, 2d, null, List.of(medicationId)))
         );
         String expectedMessage = "La medicación con id " + medicationId + " no se puede activar porque ya esta finalizada";
         String actualMessage = exception.getMessage();
@@ -176,7 +176,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medication2Id = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.STOPPED, null, 9, null)).getId();
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, (short)2, null, List.of(medication2Id)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, 2d, null, List.of(medication2Id)))
         );
         expectedMessage = "La medicación con id " + medication2Id + " no se puede activar porque ya esta finalizada";
         actualMessage = exception.getMessage();
@@ -185,7 +185,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medication3Id = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.ACTIVE, null, 9, null)).getId();
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, (short)2, null, List.of(medication3Id)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, 2d, null, List.of(medication3Id)))
         );
         expectedMessage = "La medicación con id " + medication3Id + " no se puede activar porque ya esta activa";
         actualMessage = exception.getMessage();
@@ -198,7 +198,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medication4Id = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.SUSPENDED, null, 9, dosageId)).getId();
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, (short)2, null, List.of(medication4Id)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, 2d, null, List.of(medication4Id)))
         );
         expectedMessage = "La medicación con id " + medication4Id + " no se puede activar porque ya esta activa";
         actualMessage = exception.getMessage();
@@ -217,7 +217,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medicationId = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.STOPPED, null, 9, dosageId)).getId();
 
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4),  new ChangeStateMedicationRequestBo(MedicationStatementStatus.STOPPED, (short)2, null, List.of(medicationId)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4),  new ChangeStateMedicationRequestBo(MedicationStatementStatus.STOPPED, 2d, null, List.of(medicationId)))
         );
         String expectedMessage = "La medicación con id " + medicationId + " no se puede finalizar porque ya esta finalizada";
         String actualMessage = exception.getMessage();
@@ -227,7 +227,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
         Integer medication2Id = save(MedicationTestMocks.createMedicationStatement(patientId, ibuprofenoId, "", MedicationStatementStatus.STOPPED, null, 9, null)).getId();
 
         exception = Assertions.assertThrows(IllegalArgumentException.class, () ->
-                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.STOPPED, (short)2, null, List.of(medication2Id)))
+                changeStateMedicationService.execute(new PatientInfoBo(patientId, (short) 1, (short) 4), new ChangeStateMedicationRequestBo(MedicationStatementStatus.STOPPED, 2d, null, List.of(medication2Id)))
         );
         expectedMessage = "La medicación con id " + medication2Id + " no se puede finalizar porque ya esta finalizada";
         actualMessage = exception.getMessage();
@@ -238,7 +238,7 @@ public class ChangeStateMedicationServiceImplTest extends UnitRepository {
     @Test
     public void test_execute_success(){
         PatientInfoBo patient = new PatientInfoBo(1, (short) 1, (short) 18);
-        changeStateMedicationService.execute(patient, new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, (short)2, null, Collections.emptyList()));
+        changeStateMedicationService.execute(patient, new ChangeStateMedicationRequestBo(MedicationStatementStatus.ACTIVE, 2d, null, Collections.emptyList()));
         Assertions.assertTrue(true);
     }
